Update code to jolt main branch

+++ zkml-jolt-core/src/jolt/executor/instructions/virtual_sra.rs
@@ -24,8 +24,8 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for VirtualSra {
 
     fn to_lookup_output(&self) -> u64 {
         let (x, y) = LookupQuery::<WORD_SIZE>::to_instruction_inputs(self);
-        let mut x = LookupBits::new(x, WORD_SIZE);
-        let mut y = LookupBits::new(y as u64, WORD_SIZE);
+        let mut x = LookupBits::new(x as u128, WORD_SIZE);
+        let mut y = LookupBits::new(y as u128, WORD_SIZE);
 
         let sign_bit = if x.leading_ones() == 0 { 0 } else { 1 };
         let mut entry = 0;
